Bot_Binance/requirements.txt
# Core dependencies
numpy==1.21.6
pandas==1.3.5
python-binance==1.0.16
python-telegram-bot>=20.0
websockets>=10.3
aiohttp>=3.8.3
PyYAML>=6.0
python-dotenv>=0.21.0

# GUI
tkinter  # Usually comes with Python

# Database
sqlite3  # Usually comes with Python

# Testing
pytest>=7.2.0
pytest-asyncio>=0.20.2

# Development
black>=22.10.0
flake8>=5.0.4
mypy>=0.991

# Documentation 
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.1